requirements.txt
numpy==1.26.4
transformers==4.38.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
huggingface-hub==0.21.4
python-dotenv==1.1.0
flask==3.1.0
openpyxl==3.1.5
pandas==2.2.3
requests==2.32.3
xlrd==2.0.1
pytest==8.3.5
httpx==0.28.1
pydantic==2.11.3
tqdm==4.67.1
werkzeug==3.1.3
anyio==4.9.0
jiter==0.9.0
et-xmlfile==2.0.0
safetensors==0.4.2
accelerate==0.30.1
bitsandbytes==0.43.2